refactor(native_blockifier): remove unused py objects
